added state class, fixed SODAR class, added state_test.ino

+++ arduino_code/libraries/State/State.cpp
@@ -2,12 +2,12 @@
 
 State::State()
 {
-    linear_state = 0.0;
-    rotation_state = 0.0;
+    linear_state = 0;
+    rotation_state = 0;
 }
 
-void State::setLinearState(float val) { linear_state=val;}
-void State::setRotationState(float val) { rotation_state=val;}
+void State::setLinearState(int val) { linear_state=val;}
+void State::setRotationState(int val) { rotation_state=val;}
 
-float State::getLinearState(){return linear_state;}
-float State::getRotationState(){return rotation_state;}
+int State::getLinearState(){return linear_state;}
+int State::getRotationState(){return rotation_state;}
